Rename package to specifications plural
