+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/specifications/SofortSpec.kt
@@ -1,12 +1,12 @@
-package com.stripe.android.paymentsheet.specification
+package com.stripe.android.paymentsheet.specifications
 
 import androidx.compose.ui.graphics.Color
 import com.stripe.android.paymentsheet.R
-import com.stripe.android.paymentsheet.specification.FormElementSpec.SectionSpec
-import com.stripe.android.paymentsheet.specification.FormElementSpec.SectionSpec.SectionFieldSpec.Country
-import com.stripe.android.paymentsheet.specification.FormElementSpec.SectionSpec.SectionFieldSpec.Email
-import com.stripe.android.paymentsheet.specification.FormElementSpec.SectionSpec.SectionFieldSpec.Name
-import com.stripe.android.paymentsheet.specification.FormElementSpec.StaticTextSpec
+import com.stripe.android.paymentsheet.specifications.FormElementSpec.SectionSpec
+import com.stripe.android.paymentsheet.specifications.FormElementSpec.SectionSpec.SectionFieldSpec.Country
+import com.stripe.android.paymentsheet.specifications.FormElementSpec.SectionSpec.SectionFieldSpec.Email
+import com.stripe.android.paymentsheet.specifications.FormElementSpec.SectionSpec.SectionFieldSpec.Name
+import com.stripe.android.paymentsheet.specifications.FormElementSpec.StaticTextSpec
 
 internal val sofortParams: MutableMap<String, Any?> = mutableMapOf(
     "country" to null,
